# travis ubuntu trusty to speed up tests and enable integration tests