Do not merge yet!

Build time of this PR (new travis config):
14 min 2 sec
7 min 17s

Build time with old travis config:
22 min 21 sec
26 min 40 sec
23 min 48 sec
23 min 43 sec
25 min 4 sec
22 min 26 sec