@pavolloffay 2016-08-26
@objectiser could you please review and eventually merge?